Add "New Grid" button. Change inner structure
Add new grid button that removes the previous grid and creates a new grid. If the user presses cancel
it creates a default 16x16 grid.

Also this code refactors the previous commit code by adding all of the behavior inside a createGrid function.
This function gets called at the page start and whenever the user presses the button and enters a valid prompt.